TaxWizardFInale/TaxWizard/pyproject.toml
[project]
name = "tax-wizard"
version = "0.1.0"
description = "A comprehensive tax management and calculation system"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "pypdf2>=3.0.1",
    "pdf2image>=1.17.0",
    "python-magic>=0.4.27",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "flask-wtf>=1.2.1",
    "pillow>=10.2.0",
    "requests>=2.31.0",
    "bcrypt>=4.1.2",
    "flask-migrate>=4.0.5",
    "alembic>=1.13.1",
    "flask-mail>=0.9.1",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=4.0.0",
    "flask-jwt-extended>=4.6.0",
    "flask-limiter>=3.5.0",
    "flask-caching>=2.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
